requirements.txt
flask==3.0.3
flask-cors==4.0.0
boto3==1.34.131
python-dotenv==1.0.1
requests==2.31.0
pillow==11.3.0
werkzeug==3.1.3
